split 2 sub package for base packages component scan.
